[enhancement](macOS M1) Support building from source on macOS (M1) (#13195)
# Proposed changes

This PR fixed lots of issues when building from source on macOS with Apple M1 chip.

## ATTENTION

The job for supporting macOS with Apple M1 chip is too big and there are lots of unresolved issues during runtime:
1. Some errors with memory tracker occur when BE (RELEASE) starts.
2. Some UT cases fail.
...

Temporarily, the following changes are made on macOS to start BE successfully.
1. Disable memory tracker.
2. Use tcmalloc instead of jemalloc.

This PR kicks off the job. Guys who are interested in this job can continue to fix these runtime issues.

## Use case

```shell
./build.sh -j 8 --be --clean

cd output/be/bin
ulimit -n 60000
./start_be.sh --daemon
```

## Something else

It takes around _**10+**_ minutes to build BE (with prebuilt third-parties) on macOS with M1 chip. We will improve the  development experience on macOS greatly when we finish the adaptation job.

+++ be/src/vec/aggregate_functions/aggregate_function_collect.h
@@ -84,7 +84,7 @@ struct AggregateFunctionCollectSetData<StringRef> {
         }
     }
     void read(BufferReadable& buf) {
-        size_t rows;
+        UInt64 rows;
         read_var_uint(rows, buf);
 
         StringRef ref;
@@ -121,7 +121,7 @@ struct AggregateFunctionCollectListData {
         buf.write(data.raw_data(), data.size() * sizeof(ElementType));
     }
     void read(BufferReadable& buf) {
-        size_t rows = 0;
+        UInt64 rows = 0;
         read_var_uint(rows, buf);
         data.resize(rows);
         buf.read(reinterpret_cast<char*>(data.data()), rows * sizeof(ElementType));
@@ -161,13 +161,13 @@ struct AggregateFunctionCollectListData<StringRef> {
 
     void read(BufferReadable& buf) {
         auto& col = assert_cast<ColVecType&>(*data);
-        size_t offs_size = 0;
+        UInt64 offs_size = 0;
         read_var_uint(offs_size, buf);
         col.get_offsets().resize(offs_size);
         buf.read(reinterpret_cast<char*>(col.get_offsets().data()),
                  offs_size * sizeof(IColumn::Offset));
 
-        size_t chars_size = 0;
+        UInt64 chars_size = 0;
         read_var_uint(chars_size, buf);
         col.get_chars().resize(chars_size);
         buf.read(reinterpret_cast<char*>(col.get_chars().data()), chars_size);
